NOT WORKING - partial view
Attempt at working in a 'Parital Run' Option on the Review page.

A single Map Object with a list of tiles, displayed out to the Review, with each starting with 'Partial Run?'

problem: it's not returning the list of items back from the editor.

branched to remain, but also to be able to revert.